Hotfix/bug fixes

+++ api/Share_To_Learn_WEB_API/Share_To_Learn_WEB_API/Controllers/GroupController.cs
@@ -215,12 +215,19 @@ namespace Share_To_Learn_WEB_API.Controllers
         public async Task<ActionResult> deleteGroup(int groupId)
         {
             string patha = await _repository.GetGroupImage(groupId);
-            FileManagerService.deleteFile(patha);
+            if(!string.IsNullOrEmpty(patha))
+                FileManagerService.deleteFile(patha);
 
             IEnumerable<string> paths = await _documentrepository.GetDocumentsPaths(groupId);
-
-            foreach (string path in paths)
-                FileManagerService.deleteFile(path);
+            
+            if(paths!=null)
+            { 
+                foreach (string path in paths) 
+                { 
+                    if(!string.IsNullOrEmpty(path))
+                        FileManagerService.deleteFile(path);
+                }
+            }
 
             await _repository.DeleteGroup(groupId);
